fixed double

+++ ImageData.h
@@ -4,17 +4,17 @@
 struct ImageData {
     int width = 0;
     int height = 0;
-    std::vector<std::vector<std::vector<int>>> pixels; // [height][width][RGB]
+    std::vector<std::vector<std::vector<double>>> pixels; // [height][width][RGB]
     bool isLoaded = false;
     
     void allocate(int w, int h) {
         width = w;
         height = h;
-        pixels = std::vector<std::vector<std::vector<int>>>(
+        pixels = std::vector<std::vector<std::vector<double>>>(
             height, 
-            std::vector<std::vector<int>>(
+            std::vector<std::vector<double>>(
                 width, 
-                std::vector<int>(3, 0) // RGB values
+                std::vector<double>(3, 0) // RGB values
             )
         );
         isLoaded = true;
